feat: split Source Manager from Content Manager — management via Settings, content via Research panel
Source management (add, remove, link folders) moves to a dedicated Source Manager sheet
accessed from Settings gear menu. The Sources tab in the Research panel becomes purely
content-focused: search, read, clip, and insert. This separates set-and-forget management
from every-session usage.
