+++ web/src/components/project/settings-menu.tsx
@@ -13,6 +13,8 @@ interface SettingsMenuProps {
   onUnlinkDrive?: () => void;
   /** Open Google Accounts management sheet */
   onManageAccounts?: () => void;
+  /** Open Source Manager sheet */
+  onManageSources?: () => void;
   /** Open rename book dialog */
   onRenameBook: () => void;
   /** Open duplicate book dialog */
@@ -42,6 +44,7 @@ export function SettingsMenu({
   onSetupDrive,
   onUnlinkDrive,
   onManageAccounts,
+  onManageSources,
   onRenameBook,
   onDuplicateBook,
   isDuplicating,
@@ -217,6 +220,31 @@ export function SettingsMenu({
                 />
               </svg>
               Google Accounts
+            </button>
+          )}
+
+          {/* Source Manager */}
+          {onManageSources && (
+            <button
+              onClick={() => handleMenuItem(onManageSources)}
+              className="w-full text-left px-4 py-2.5 text-sm text-gray-700 hover:bg-gray-100
+                         transition-colors min-h-[44px] flex items-center gap-2"
+              role="menuitem"
+            >
+              <svg
+                className="w-4 h-4 shrink-0"
+                fill="none"
+                stroke="currentColor"
+                viewBox="0 0 24 24"
+              >
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth={2}
+                  d="M19 11H5m14 0a2 2 0 012 2v6a2 2 0 01-2 2H5a2 2 0 01-2-2v-6a2 2 0 012-2m14 0V9a2 2 0 00-2-2M5 11V9a2 2 0 012-2m0 0V5a2 2 0 012-2h6a2 2 0 012 2v2M7 7h10"
+                />
+              </svg>
+              Manage Sources
             </button>
           )}
