+++ web/src/components/research/research-panel-provider.tsx
@@ -5,21 +5,16 @@ import { createContext, useContext, useReducer, useCallback, useMemo, type React
 // === State Types ===
 
 export type ResearchTab = "sources" | "ask" | "clips";
-export type SourcesView = "list" | "detail" | "add-document";
-export type AddMode = "link-folder" | "add-documents" | null;
+export type SourcesView = "list" | "detail";
 
 export interface ResearchPanelState {
   isOpen: boolean;
   activeTab: ResearchTab;
   sourcesView: SourcesView;
   activeSourceId: string | null;
-  /** Active Drive connection ID — used by add-document flow for pre-selection */
-  activeConnectionId: string | null;
   returnTab: "ask" | "clips" | null;
   /** Text to scroll to when viewing a source detail (from citation navigation) */
   scrollToText: string | null;
-  /** Distinguishes link-folder from add-documents in the add-document flow */
-  addMode: AddMode;
 }
 
 // === Actions ===
@@ -30,11 +25,7 @@ export type ResearchPanelAction =
   | { type: "SET_TAB"; tab: ResearchTab }
   | { type: "VIEW_SOURCE"; sourceId: string; returnTo?: "ask" | "clips"; scrollToText?: string }
   | { type: "BACK_TO_LIST" }
-  | { type: "RETURN_TO_TAB" }
-  | { type: "START_ADD_DOCUMENT"; connectionId?: string }
-  | { type: "START_LINK_FOLDER"; connectionId?: string }
-  | { type: "SET_DRIVE_CONNECTION"; connectionId: string }
-  | { type: "FINISH_FLOW" };
+  | { type: "RETURN_TO_TAB" };
 
 // === Initial State ===
 
@@ -43,10 +34,8 @@ const initialState: ResearchPanelState = {
   activeTab: "sources",
   sourcesView: "list",
   activeSourceId: null,
-  activeConnectionId: null,
   returnTab: null,
   scrollToText: null,
-  addMode: null,
 };
 
 // === Reducer ===
@@ -54,14 +43,14 @@ const initialState: ResearchPanelState = {
 /**
  * State machine reducer for the Research Panel.
  *
- * Valid states (6):
+ * Valid states (5):
  *   1. Closed              - panel is not visible
- *   2. Sources-List        - panel open, sources tab, two-section list view
+ *   2. Sources-List        - panel open, sources tab, flat list view
  *   3. Sources-Detail      - panel open, sources tab, viewing a specific source
- *   4. Sources-AddDocument - panel open, sources tab, add document flow (drive browser / upload)
- *   5. Ask                 - panel open, ask tab
- *   6. Clips               - panel open, clips tab
+ *   4. Ask                 - panel open, ask tab
+ *   5. Clips               - panel open, clips tab
  *
+ * Note: Add-document flow was moved to the Source Manager sheet (separate component).
  * Invalid transitions are prevented by returning current state unchanged.
  */
 export function researchPanelReducer(
@@ -118,18 +107,16 @@ export function researchPanelReducer(
     }
 
     case "BACK_TO_LIST": {
-      // Valid from detail or add-document views
+      // Valid from detail view only
       if (!state.isOpen || state.activeTab !== "sources") return state;
-      if (state.sourcesView !== "detail" && state.sourcesView !== "add-document") return state;
+      if (state.sourcesView !== "detail") return state;
 
       return {
         ...state,
         sourcesView: "list",
         activeSourceId: null,
-        activeConnectionId: null,
         returnTab: null,
         scrollToText: null,
-        addMode: null,
       };
     }
 
@@ -148,59 +135,6 @@ export function researchPanelReducer(
       };
     }
 
-    case "START_ADD_DOCUMENT": {
-      // Can only start add document flow when panel is open
-      if (!state.isOpen) return state;
-
-      return {
-        ...state,
-        activeTab: "sources",
-        sourcesView: "add-document",
-        activeSourceId: null,
-        activeConnectionId: action.connectionId ?? null,
-        returnTab: null,
-        addMode: "add-documents",
-      };
-    }
-
-    case "START_LINK_FOLDER": {
-      // Can only start link folder flow when panel is open
-      if (!state.isOpen) return state;
-
-      return {
-        ...state,
-        activeTab: "sources",
-        sourcesView: "add-document",
-        activeSourceId: null,
-        activeConnectionId: action.connectionId ?? null,
-        returnTab: null,
-        addMode: "link-folder",
-      };
-    }
-
-    case "SET_DRIVE_CONNECTION": {
-      // Only valid during add-document flow
-      if (!state.isOpen || state.sourcesView !== "add-document") return state;
-
-      return {
-        ...state,
-        activeConnectionId: action.connectionId,
-      };
-    }
-
-    case "FINISH_FLOW": {
-      // Valid during add-document flow
-      if (!state.isOpen) return state;
-      if (state.sourcesView !== "add-document") return state;
-
-      return {
-        ...state,
-        sourcesView: "list",
-        activeConnectionId: null,
-        addMode: null,
-      };
-    }
-
     default:
       return state;
   }
@@ -214,10 +148,8 @@ export interface ResearchPanelContextValue {
   activeTab: ResearchTab;
   sourcesView: SourcesView;
   activeSourceId: string | null;
-  activeConnectionId: string | null;
   returnTab: "ask" | "clips" | null;
   scrollToText: string | null;
-  addMode: AddMode;
 
   // Actions
   openPanel: (tab?: ResearchTab) => void;
@@ -226,10 +158,6 @@ export interface ResearchPanelContextValue {
   viewSource: (sourceId: string, returnTo?: "ask" | "clips", scrollToText?: string) => void;
   backToSourceList: () => void;
   returnToPreviousTab: () => void;
-  startAddDocument: (connectionId?: string) => void;
-  startLinkFolder: (connectionId?: string) => void;
-  setDriveConnection: (connectionId: string) => void;
-  finishFlow: () => void;
 
   // Raw dispatch for testing or advanced usage
   dispatch: React.Dispatch<ResearchPanelAction>;
@@ -279,31 +207,6 @@ export function ResearchPanelProvider({ children }: ResearchPanelProviderProps) 
     dispatch({ type: "RETURN_TO_TAB" });
   }, [dispatch]);
 
-  const startAddDocument = useCallback(
-    (connectionId?: string) => {
-      dispatch({ type: "START_ADD_DOCUMENT", connectionId });
-    },
-    [dispatch],
-  );
-
-  const startLinkFolder = useCallback(
-    (connectionId?: string) => {
-      dispatch({ type: "START_LINK_FOLDER", connectionId });
-    },
-    [dispatch],
-  );
-
-  const setDriveConnection = useCallback(
-    (connectionId: string) => {
-      dispatch({ type: "SET_DRIVE_CONNECTION", connectionId });
-    },
-    [dispatch],
-  );
-
-  const finishFlow = useCallback(() => {
-    dispatch({ type: "FINISH_FLOW" });
-  }, [dispatch]);
-
   const value = useMemo<ResearchPanelContextValue>(
     () => ({
       // State
@@ -311,10 +214,8 @@ export function ResearchPanelProvider({ children }: ResearchPanelProviderProps) 
       activeTab: state.activeTab,
       sourcesView: state.sourcesView,
       activeSourceId: state.activeSourceId,
-      activeConnectionId: state.activeConnectionId,
       returnTab: state.returnTab,
       scrollToText: state.scrollToText,
-      addMode: state.addMode,
 
       // Actions
       openPanel,
@@ -323,27 +224,11 @@ export function ResearchPanelProvider({ children }: ResearchPanelProviderProps) 
       viewSource,
       backToSourceList,
       returnToPreviousTab,
-      startAddDocument,
-      startLinkFolder,
-      setDriveConnection,
-      finishFlow,
 
       // Raw dispatch
       dispatch,
     }),
-    [
-      state,
-      openPanel,
-      closePanel,
-      setActiveTab,
-      viewSource,
-      backToSourceList,
-      returnToPreviousTab,
-      startAddDocument,
-      startLinkFolder,
-      setDriveConnection,
-      finishFlow,
-    ],
+    [state, openPanel, closePanel, setActiveTab, viewSource, backToSourceList, returnToPreviousTab],
   );
 
   return <ResearchPanelContext.Provider value={value}>{children}</ResearchPanelContext.Provider>;
